requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
dataframe-image==0.1.1
defusedxml==0.7.1
entrypoints==0.4
fastjsonschema==2.15.3
filelock==3.7.1
fonttools==4.33.3
frozenlist==1.3.0
idna==3.3
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.6.0
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.2
langdetect==1.0.9
lxml==4.9.0
MarkupSafe==2.1.1
matplotlib==3.5.2
mistune==0.8.4
multidict==6.0.2
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
Pillow==9.1.1
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pyzmq==23.1.0
regex==2022.6.2
requests==2.27.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
snscrape==0.4.3.20220106
soupsieve==2.3.2.post1
textblob==0.17.1
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
urllib3==1.26.9
webencodings==0.5.1
yarl==1.7.2
